requirements_cpu.txt
GitPython~=3.1
numpy~=1.20.3
tqdm<4.50.0,>=4.27
torch==1.8.1+cpu
torchtext<0.10
transformers~=4.8.2
spacy~=3.0.6
scikit-learn~=0.24.2
pandas~=1.2.5
datasets~=1.8.0
coloredlogs~=15.0.1
hyperopt~=0.2.5
yapf
emoji
sentencepiece
dill
clean-text
neuspell
unidecode
nltk
modin[dask]

# Viz in notebook
matplotlib
mpld3
mplcursors
wordcloud

# package location
--find-links https://download.pytorch.org/whl/lts/1.8/torch_lts.html
